requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
dash
gunicorn==23.0.0
